+++ cli/src/commands/file/mod.rs
@@ -15,6 +15,7 @@
 pub mod chmod;
 pub mod list;
 pub mod show;
+pub mod untrack;
 
 use crate::cli_util::CommandHelper;
 use crate::command_error::CommandError;
@@ -26,6 +27,7 @@ pub enum FileCommand {
     Chmod(chmod::FileChmodArgs),
     List(list::FileListArgs),
     Show(show::FileShowArgs),
+    Untrack(untrack::FileUntrackArgs),
 }
 
 pub fn cmd_file(
@@ -37,5 +39,6 @@ pub fn cmd_file(
         FileCommand::Chmod(args) => chmod::cmd_file_chmod(ui, command, args),
         FileCommand::List(args) => list::cmd_file_list(ui, command, args),
         FileCommand::Show(args) => show::cmd_file_show(ui, command, args),
+        FileCommand::Untrack(args) => untrack::cmd_file_untrack(ui, command, args),
     }
 }