+++ cli/src/commands/file/untrack.rs
@@ -28,7 +28,7 @@ use crate::ui::Ui;
 
 /// Stop tracking specified paths in the working copy
 #[derive(clap::Args, Clone, Debug)]
-pub(crate) struct UntrackArgs {
+pub(crate) struct FileUntrackArgs {
     /// Paths to untrack. They must already be ignored.
     ///
     /// The paths could be ignored via a .gitignore or .git/info/exclude (in
@@ -38,10 +38,10 @@ pub(crate) struct UntrackArgs {
 }
 
 #[instrument(skip_all)]
-pub(crate) fn cmd_untrack(
+pub(crate) fn cmd_file_untrack(
     ui: &mut Ui,
     command: &CommandHelper,
-    args: &UntrackArgs,
+    args: &FileUntrackArgs,
 ) -> Result<(), CommandError> {
     let mut workspace_command = command.workspace_helper(ui)?;
     let store = workspace_command.repo().store().clone();